+++ be/src/exec/hash-table.inline.h
@@ -11,10 +11,10 @@ inline HashTable::Iterator HashTable::Find(TupleRow* probe_row) {
   bool has_nulls = EvalProbeRow(probe_row);
   if (!stores_nulls_ && has_nulls) return End();
   uint32_t hash = HashCurrentRow();
-  int bucket_idx = hash % num_buckets_;
+  int64_t bucket_idx = hash % num_buckets_;
 
   Bucket* bucket = &buckets_[bucket_idx];
-  int node_idx = bucket->node_idx_;
+  int64_t node_idx = bucket->node_idx_;
   while (node_idx != -1) {
     Node* node = GetNode(node_idx);
     if (node->hash_ == hash && Equals(node->data())) {
@@ -27,7 +27,7 @@ inline HashTable::Iterator HashTable::Find(TupleRow* probe_row) {
 }
   
 inline HashTable::Iterator HashTable::Begin() {
-  int bucket_idx = -1;
+  int64_t bucket_idx = -1;
   Bucket* bucket = NextBucket(&bucket_idx);
   if (bucket != NULL) {
     return Iterator(this, bucket_idx, bucket->node_idx_, 0);
@@ -35,7 +35,7 @@ inline HashTable::Iterator HashTable::Begin() {
   return End();
 }
 
-inline HashTable::Bucket* HashTable::NextBucket(int* bucket_idx) {
+inline HashTable::Bucket* HashTable::NextBucket(int64_t* bucket_idx) {
   ++*bucket_idx;
   for (; *bucket_idx < num_buckets_; ++*bucket_idx) {
     if (buckets_[*bucket_idx].node_idx_ != -1) return &buckets_[*bucket_idx];
@@ -49,7 +49,7 @@ inline void HashTable::InsertImpl(TupleRow* row) {
   if (!stores_nulls_ && has_null) return;
 
   uint32_t hash = HashCurrentRow();
-  int bucket_idx = hash % num_buckets_;
+  int64_t bucket_idx = hash % num_buckets_;
   if (num_nodes_ == nodes_capacity_) GrowNodeArray();
   Node* node = GetNode(num_nodes_);
   TupleRow* data = node->data();
@@ -76,7 +76,7 @@ inline void HashTable::Iterator::Next() {
   // for the current probe row.
   if (check_match) {
     // TODO: this should prefetch the next node
-    int next_idx = node->next_idx_;
+    int64_t next_idx = node->next_idx_;
     while (next_idx != -1) {
       node = table_->GetNode(next_idx);
       if (node->hash_ == scan_hash_ && table_->Equals(node->data())) {
